业务逻辑加而且判断

+++ src/main/java/com/bcx/plat/core/controller/FrontFuncController.java
@@ -1,12 +1,15 @@
 package com.bcx.plat.core.controller;
 
+import com.bcx.plat.core.base.BaseConstants;
 import com.bcx.plat.core.common.BaseControllerTemplate;
+import com.bcx.plat.core.constants.Message;
 import com.bcx.plat.core.entity.FrontFunc;
 import com.bcx.plat.core.morebatis.component.FieldCondition;
 import com.bcx.plat.core.morebatis.component.constant.Operator;
 import com.bcx.plat.core.service.BusinessObjectService;
 import com.bcx.plat.core.service.FrontFuncProService;
 import com.bcx.plat.core.service.FrontFuncService;
+import com.bcx.plat.core.utils.PlatResult;
 import com.bcx.plat.core.utils.ServiceResult;
 import com.bcx.plat.core.utils.UtilsTool;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -79,15 +82,18 @@ public class FrontFuncController extends BaseControllerTemplate<FrontFuncService
     @RequestMapping("/delete")
     @Override
     public Object delete(String rowId, HttpServletRequest request, Locale locale) {
-        List<Map<String, Object>> list = frontFuncProService
-                .select(new FieldCondition("funcRowId", Operator.EQUAL, rowId));
-        if (UtilsTool.isValid(list)) {
-            List<String> rowIds = list.stream().map((row) -> {
-                return (String) row.get("rowId");
-            }).collect(Collectors.toList());
-            frontFuncProService.delete(new FieldCondition("rowId", Operator.IN, rowIds));
+        if (UtilsTool.isValid(rowId)) {
+            List<Map<String, Object>> list = frontFuncProService
+                    .select(new FieldCondition("funcRowId", Operator.EQUAL, rowId));
+            if (UtilsTool.isValid(list)) {
+                List<String> rowIds = list.stream().map((row) -> {
+                    return (String) row.get("rowId");
+                }).collect(Collectors.toList());
+                frontFuncProService.delete(new FieldCondition("rowId", Operator.IN, rowIds));
+            }
+            return super.delete(rowId, request, locale);
         }
-        return super.delete(rowId, request, locale);
+        return super.result(request,ServiceResult.Msg(PlatResult.Msg(BaseConstants.STATUS_FAIL, Message.QUERY_FAIL)),locale);
     }
 
 
@@ -101,7 +107,10 @@ public class FrontFuncController extends BaseControllerTemplate<FrontFuncService
      */
     @RequestMapping("/queryFuncCode")
     public Object queryFuncCode(String funcCode, HttpServletRequest request, Locale locale) {
-        List list = UtilsTool.jsonToObj(funcCode.toString(), List.class);
-        return super.result(request, ServiceResult.Msg(frontFuncService.queryFuncCode(list)),locale);
+        if (UtilsTool.isValid(funcCode)) {
+            List list = UtilsTool.jsonToObj(funcCode, List.class);
+            return super.result(request, ServiceResult.Msg(frontFuncService.queryFuncCode(list)),locale);
+        }
+        return super.result(request, ServiceResult.Msg(PlatResult.Msg(BaseConstants.STATUS_FAIL,Message.QUERY_FAIL)),locale);
     }
 }
